requirements.txt
opencv-python
mediapipe
numpy
Pillow
scikit-learn
pandas
